Remove unnecessary registry field definition
This field definition is only needed if no interface is given. By
default it uses the definition given by the interface.

This field definition also lead to some problems when upgrading from
Plone 4 to 5. My guess is that the definition or internals of the schema
changed to some degree. Without the field definition in the registry.xml
it works fine for new installations and upgrades.