Fix reinstallation of the package if it was initially installed on 1.1.0
__Problem__

In the release 1.1.0 and 1.1.1 ftw.datepicker added registry entries as
follows:
```xml
<record interface="ftw.datepicker.interfaces.IDatetimeRegistry" field="formats">
    <value purge="false">
        <element key="de">d.m.Y H:i</element>
        <element key="fr">d/m/Y H:i</element>
    </value>
</record>
```

Like the created record in the registry has a field which knows about the
interface (thanks to the `interface="..."`). If a field knows about the
`interface` the record will provide IInterfaceAwareRecord __on
initialization__.

In version 1.2.0 the record declaration was changed to:
```xml
<record name="ftw.datepicker.interfaces.IDatetimeRegistry.formats">
    <field type="plone.registry.field.Dict">
        <title>The formats per language</title>
        <key_type type="plone.registry.field.TextLine" />
        <value_type type="plone.registry.field.TextLine" />
    </field>
    <value purge="false">
        <element key="de">d.m.Y H:i</element>
        <element key="fr">d/m/Y H:i</element>
    </value>
</record>
```

In here we do not define the interface anymore but do define the field
itself.

Now when you have initially installed version 1.1.0 or 1.1.1 the field
knows the interface. When you reinstall the profile with version 1.2.0
and above plone.app.registry will get the existing record from the
registry. In that moment it sees that it's field knows about the
interface and thus provides the IInterfaceAwareRecord interface. During
the registration of the entry it finds that we declare the `field`
manually in the XML. Because of that it creates a new field in case
something has changed. During this creation process of the field it lets
the field know about the interface. __BUT__ because we didn't define an
interface this time in the record XML-node it sets `None` instead of our
actual interface.

Now during the rest of the installation process the record provides
IInterfaceAwareRecord even though the field does __not__ know about the
interface anymore.

Later on in the record registration process this will trigger the
`redispatchInterfaceAwareRecordEvents` subscriber from `plone.registry`.
In that moment he thinks the record knows about the interface because it
still provides IInterfaceAwareRecord. This causes a failure when trying
to resolve that interface.

---

__Solution__

Our solution is to remove the interface from the field before we
reinstall the package. So that during the installation it never sets the
IInterfaceAwareRecord interface.